# docs/api-spec.yml
openapi: '3.0.3'
info:
  title: KillBug API
  description: Auto-generated by api-spec-update.py
  version: '1.0.0'
servers:
  - url: http://localhost:8080
    description: webhook-server
paths:
  /health:
    get:
      operationId: WebhookController_getHealth
      summary: WebhookController.getHealth
      responses:
        '200':
          description: OK
    get:
      operationId: LinearWebhookController_getHealth
      summary: LinearWebhookController.getHealth
      responses:
        '200':
          description: OK
  /linear/webhook:
    post:
      operationId: WebhookController_handleLinearWebhook
      summary: WebhookController.handleLinearWebhook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: OK
    post:
      operationId: LinearWebhookController_handleLinearWebhook
      summary: LinearWebhookController.handleLinearWebhook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: OK
  /role-data/{role}:
    get:
      operationId: RoleDataController_handleList
      summary: RoleDataController.handleList
      parameters:
        - name: role
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleDataResponse'
    post:
      operationId: RoleDataController_handleCreate
      summary: RoleDataController.handleCreate
      parameters:
        - name: role
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoleDataCreateRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleDataResponse'
  /senior-reviews:
    get:
      operationId: SeniorReviewController_handleSearch
      summary: SeniorReviewController.handleSearch
      parameters:
        - name: module
          in: query
          required: false
          schema:
            type: string
        - name: name
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SeniorReviewResponse'
  /slack/commands:
    post:
      operationId: SlackCommandController_handleCommand
      summary: SlackCommandController.handleCommand
      parameters:
        - name: X-Slack-Request-Timestamp
          in: header
          required: false
          schema:
            type: string
        - name: X-Slack-Signature
          in: header
          required: false
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
      responses:
        '200':
          description: OK
  /slack/events:
    post:
      operationId: SlackController_handleSlackEvent
      summary: SlackController.handleSlackEvent
      parameters:
        - name: X-Slack-Request-Timestamp
          in: header
          required: false
          schema:
            type: string
        - name: X-Slack-Signature
          in: header
          required: false
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: OK
components:
  schemas:
    RoleDataCreateRequest:
      type: object
      properties:
        name:
          type: string
        content:
          type: string
        extra:
          type: string
    RoleDataResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        content:
          type: string
        extra:
          type: string
        createdAt:
          type: string
          format: date-time
    SeniorReviewResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        review:
          type: string
        module:
          type: string
        createdAt:
          type: string
          format: date-time